+++ CHANGELOG.md
@@ -1,4 +1,15 @@
-# Working Log
+# Change Log
+
+### 1.0.2 - 2022-01-03
+
+#### Added
+
+-   Divide tool by area. 
+
+#### Fixed
+
+-   Divide using distance. [1](https://github.com/3DGISKing/QGISEditTools/issues/1)
+-   Divide using area. [1](https://github.com/3DGISKing/QGISEditTools/issues/2)
 
 ### 1.0.1 - 2021-12-21
 
@@ -15,7 +26,7 @@
 
 #### Added
 
--   Divide Tool.
+-   Divide Tool by distance.
 
 #### Added